Added unit tests for server.cc
Session class needed to be mocked to properly add unit tests for server.cc. This resulted in needing to create an interface for session to allow for easier mocking of the class.

Additionally needed to change the constructor for server to take a session variable so that the functions for session could be mocked.

Updated server_main to reflect these changes.

Change-Id: I4f7855b18c1a9502fbd3b6f02d70e5d1386a4ef3